[#2942] fix bug for osx

+++ nes-benchmark/include/DataGeneration/YSBDataGenerator.hpp
@@ -24,7 +24,7 @@ class YSBDataGenerator : public DataGenerator {
 
     std::string getName() override;
 
-    std::vector<Runtime::TupleBuffer> createData(uint64_t numberOfBuffers, uint64_t bufferSize) override;
+    std::vector<Runtime::TupleBuffer> createData(size_t numberOfBuffers, size_t bufferSize) override;
 
     SchemaPtr getSchema() override;
